+++ src/components/Dropdown/components/menu/menu.jsx
@@ -5,11 +5,11 @@ import "./menu.scss";
 
 const Menu = props => {
   const { children, Renderer, selectProps } = props;
-  const insideScroll = selectProps?.selectProps?.insideScroll;
+  const withFixedPosition = selectProps?.selectProps?.insideOverflowContainer;
   return (
     <components.Menu
       {...props}
-      className={cx("menu", "dropdown-menu-wrapper", { ["dropdown-menu-wrapper--inside-scroll"]: insideScroll })}
+      className={cx("menu", "dropdown-menu-wrapper", { ["dropdown-menu-wrapper--fixed-position"]: withFixedPosition })}
     >
       {Renderer && Renderer(props)}
       {!Renderer && children}